# server/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SDK Playground API",
    "version": "1.0.0",
    "description": "Auto-generated OpenAPI spec for the SDK endpoints (authenticate, createTransaction, getTransaction, refundTransaction, cancelTransaction)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/authenticate": {
      "post": {
        "summary": "Authenticate to obtain a Bearer token",
        "description": "Calls the SDK authenticate function. Provide username and password to receive a merchant token and expires_in.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "username": { "type": "string" },
                  "password": { "type": "string" }
                },
                "required": [ "username", "password" ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/ModoMerchant" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/createTransaction": {
      "post": {
        "summary": "Create a payment request (createTransaction)",
        "description": "Create a payment request using the provided token and payment payload. The token must be included in the request body as `token`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "description": { "type": "string" },
                  "amount": { "type": "number", "format": "float" },
                  "currency": { "type": "string" },
                  "cc_code": { "type": "string" },
                  "processor_code": { "type": "string" },
                  "external_intention_id": { "type": "string" },
                  "expiration_date": { "type": "string", "format": "date-time" },
                  "message": { "type": "string" },
                  "webhook_notification": { "type": "string" },
                  "customer": { "type": "object" },
                  "shipping_address": { "type": "object" },
                  "items": { "type": "array", "items": { "type": "object" } },
                  "establishment_numbers": { "type": "array", "items": { "type": "string" } }
                },
                "required": [ "token", "description", "amount", "currency", "cc_code", "processor_code", "external_intention_id" ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment request created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/CreatePaymentResponse" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/getTransaction": {
      "post": {
        "summary": "Get payment request data",
        "description": "Retrieve the payment request and transaction data for a given paymentRequestId. Requires token and paymentRequestId in the body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "paymentRequestId": { "type": "string" }
                },
                "required": [ "token", "paymentRequestId" ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Payment request data retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/GetPaymentResponse" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/refundTransaction": {
      "post": {
        "summary": "Refund a payment (partial or full)",
        "description": "Request a refund for a paymentRequestId. If amount is omitted a full refund is requested. Provide token and paymentRequestId in the body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "paymentRequestId": { "type": "string" },
                  "amount": { "type": "number", "format": "float" },
                  "body": { "type": "object", "properties": { "amount": { "type": "number" } } }
                },
                "required": [ "token", "paymentRequestId" ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Refund processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/RefundResponse" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/cancelTransaction": {
      "post": {
        "summary": "Cancel a payment (implemented as full refund)",
        "description": "Cancel a payment request. This SDK implements cancel as a full refund call to the refund endpoint. Provide token and paymentRequestId in the body.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "token": { "type": "string" },
                  "paymentRequestId": { "type": "string" }
                },
                "required": [ "token", "paymentRequestId" ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cancellation (refund) processed",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": { "type": "boolean", "example": true },
                    "data": { "$ref": "#/components/schemas/RefundResponse" }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ModoMerchant": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "expires_in": { "type": "number" }
        }
      },
      "CreatePaymentResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "qr": { "type": "string" },
          "deeplink": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "expiration_at": { "type": "string", "format": "date-time" },
          "expiration_date": { "type": "string", "format": "date-time" },
          "sub_payments": { "type": "array", "items": { "type": "object" } },
          "best_installment": { "type": "object" }
        }
      },
      "GetPaymentResponse": {
        "type": "object",
        "properties": {
          "payment_request": { "$ref": "#/components/schemas/CreatePaymentResponse" },
          "transaction_data": { "type": "object" }
        }
      },
      "RefundResponse": {
        "type": "object",
        "properties": {
          "amount": { "type": "number" },
          "status": { "type": "string" },
          "reference_transaction_token": { "type": "string" },
          "currency": { "type": "string" },
          "message": { "type": "string" }
        }
      }
    }
  }
}